ConditionSprite + DamageCondition
Rare Description Moment:
The condition sprite is the sprite that appears under an enemy when afflicted with some form of status debuff.

Also, we now have damage conditions, working in a similar way to action conditions.

Also fixed unintended logic that wasn't applying critical damage to enemies with matching weaknesses.